# scripts/reddit-openapi-complete.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Reddit API",
    "description": "Generated from Reddit JSON endpoints",
    "version": "1.0.0",
    "license": {
      "name": "Reddit API Terms",
      "url": "https://www.redditinc.com/policies/data-api-terms"
    }
  },
  "servers": [
    {
      "url": "https://www.reddit.com",
      "description": "Reddit API server"
    }
  ],
...
  "paths": {
...
    "/subreddits/popular.json": {
      "get": {
        "operationId": "getPopularSubreddits",
        "summary": "Get popular subreddits",
        "tags": ["subreddits"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPopularSubredditsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}